# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Enterprise HRMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leave-balances:
    get:
      tags: [leave]
      summary: List leave balances
      operationId: listLeaveBalances
      parameters:
        - name: employeeId
          in: query
          schema: { type: number }
        - name: year
          in: query
          schema: { type: number }
        - name: leaveTypeId
          in: query
          schema: { type: number }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LeaveBalance" }
    post:
      tags: [leave]
      summary: Create or upsert leave balance
      operationId: createLeaveBalance
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeaveBalanceInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeaveBalance" }
  /leave-balances/{id}:
    patch:
      tags: [leave]
      summary: Adjust leave balance
      operationId: updateLeaveBalance
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeaveBalanceInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeaveBalance" }

  # ── Leave Requests ────────────────────────────────────────────────────────
...
  /leave-delegations:
    get:
      tags: [leave]
      summary: List leave delegations
      operationId: listLeaveDelegations
      parameters:
        - name: delegatorId
          in: query
          schema: { type: number }
        - name: activeOnly
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LeaveDelegation" }
    post:
      tags: [leave]
      summary: Create delegation
      operationId: createLeaveDelegation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeaveDelegationInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeaveDelegation" }
  /leave-delegations/{id}:
    patch:
      tags: [leave]
      summary: Update delegation
      operationId: updateLeaveDelegation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeaveDelegationInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeaveDelegation" }
    delete:
      tags: [leave]
      summary: Delete delegation
      operationId: deleteLeaveDelegation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: No Content

  # ── Public Holidays ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    LeaveBalance:
      type: object
      properties:
        id: { type: number }
        employeeId: { type: number }
        leaveTypeId: { type: number }
        year: { type: number }
        openingBalance: { type: string }
        accrued: { type: string }
        used: { type: string }
        pending: { type: string }
        adjustment: { type: string }
        carriedOver: { type: string }
        available: { type: string }
        employeeNameEn: { type: string }
        employeeNameAr: { type: string }
        leaveTypeNameEn: { type: string }
        leaveTypeNameAr: { type: string }
        leaveTypeColor: { type: string }
      required: [id, employeeId, leaveTypeId, year, openingBalance, accrued, used, pending, adjustment, carriedOver]
    LeaveBalanceInput:
      type: object
      properties:
        employeeId: { type: number }
        leaveTypeId: { type: number }
        year: { type: number }
        openingBalance: { type: string }
        accrued: { type: string }
        adjustment: { type: string }
        carriedOver: { type: string }
      required: [employeeId, leaveTypeId, year]

    # ── Leave Requests ────────────────────────────────────────
...
    LeaveDelegation:
      type: object
      properties:
        id: { type: number }
        delegatorEmployeeId: { type: number }
        delegateeEmployeeId: { type: number }
        delegatorNameEn: { type: string }
        delegateeNameEn: { type: string }
        startDate: { type: string }
        endDate: { type: string }
        reason: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string }
      required: [id, delegatorEmployeeId, delegateeEmployeeId, startDate, endDate, isActive, createdAt]
    LeaveDelegationInput:
      type: object
      properties:
        delegatorEmployeeId: { type: number }
        delegateeEmployeeId: { type: number }
        startDate: { type: string }
        endDate: { type: string }
        reason: { type: ["string", "null"] }
        isActive: { type: boolean }
      required: [delegatorEmployeeId, delegateeEmployeeId, startDate, endDate]

    # ── Public Holidays ───────────────────────────────────────
...