# lib/api-zod/src/generated/types/leaveBalanceInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Enterprise HRMS API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaveBalanceInput {
  employeeId: number;
  leaveTypeId: number;
  year: number;
  openingBalance?: string;
  accrued?: string;
  adjustment?: string;
  carriedOver?: string;
}