# lib/api-zod/src/generated/types/leaveDelegation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Enterprise HRMS API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaveDelegation {
  id: number;
  delegatorEmployeeId: number;
  delegateeEmployeeId: number;
  delegatorNameEn?: string;
  delegateeNameEn?: string;
  startDate: string;
  endDate: string;
  /** @nullable */
  reason?: string | null;
  isActive: boolean;
  createdAt: string;
}